move sparse_util to magnification
